Count frequency of each character
Given a string that contains only lowercase characters. Write a program to print all the distinct characters along with their frequency in the order of their occurrence. For example if the string is “helloworld”, then you should print h1 e1 l3 o2 w1 r1 d1

Complete the given function countFrequency() and print the frequency of each character as per the given statement.