Pair with the given sum
Given a sorted array of n elements, and an integer k. Find whether there exists any two elements in the array that sums up to the given value k. If such a pair is found, then print “Found”, else print “Not Found”.

You may not use the same element twice.